[LEXER][Rename Interface and fix norm errors]
